requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-request==1.5.6
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
dnspython==2.1.0
idna==3.2
Pillow==8.3.1
py3dns==3.2.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-magic-bin==0.4.14
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
validate-email==1.3
